WINTERMUTE: Implement load from GMM.

+++ engines/wintermute/base/base_persistence_manager.cpp
@@ -89,7 +89,7 @@ BasePersistenceManager::BasePersistenceManager(const char *savePrefix, bool dele
 //////////////////////////////////////////////////////////////////////////
 BasePersistenceManager::~BasePersistenceManager() {
 	cleanup();
-	if (_deleteSingleton)
+	if (_deleteSingleton && BaseEngine::instance().getGameRef() == NULL)
 		BaseEngine::destroy();
 }
